# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Masafah API Documentation",
        "description": "Api description...",
        "version": "1.0.0"
    },
    "paths": {
...
        "/~tvavisa/masafah/public/api/user/addAddress": {
            "post": {
                "tags": [
                    "User Address"
                ],
                "summary": "Add User address",
                "operationId": "addAddress",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "user prefered language",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "user access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Version",
                        "in": "header",
                        "description": "1.0.0",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "description": "Address name",
                                    "type": "string",
                                    "example": "Home, office"
                                },
                                "block": {
                                    "description": "Block",
                                    "type": "string",
                                    "example": "12, 13B"
                                },
                                "street": {
                                    "description": "Street",
                                    "type": "string",
                                    "example": "12, 14A"
                                },
                                "area": {
                                    "description": "Area",
                                    "type": "string",
                                    "example": "Salmiya, Sharq"
                                },
                                "building": {
                                    "description": "Building number",
                                    "type": "string",
                                    "example": "14, 13Z"
                                },
                                "details": {
                                    "description": "Any other address details",
                                    "type": "string",
                                    "example": "Al-Kuwait, near Hamra"
                                },
                                "notes": {
                                    "description": "Extra user notes",
                                    "type": "string",
                                    "example": "Do not park vehicle infront of the gate"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    }
                }
            }
        },
        "/~tvavisa/masafah/public/api/user/getAddressById/{id}": {
            "get": {
                "tags": [
                    "User Address"
                ],
                "summary": "Get User address by ID",
                "operationId": "getAddress",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "user prefered language",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "user access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Version",
                        "in": "header",
                        "description": "1.0.0",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Address ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "404": {
                        "description": "Address not found"
                    }
                }
            }
        },
        "/~tvavisa/masafah/public/api/user/getAddresses": {
            "get": {
                "tags": [
                    "User Address"
                ],
                "summary": "Get all addresses of a user",
                "operationId": "getAddresses",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "user prefered language",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "user access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Version",
                        "in": "header",
                        "description": "1.0.0",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    }
                }
            }
        },
        "/~tvavisa/masafah/public/api/user/editAddress": {
            "put": {
                "tags": [
                    "User Address"
                ],
                "summary": "Edit Address",
                "operationId": "editAddress",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "user prefered language",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "user access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Version",
                        "in": "header",
                        "description": "1.0.0",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "address_id": {
                                    "description": "Address ID",
                                    "type": "integer",
                                    "example": 24
                                },
                                "name": {
                                    "description": "Address name",
                                    "type": "string",
                                    "example": "Home, office"
                                },
                                "block": {
                                    "description": "Block",
                                    "type": "string",
                                    "example": "12, 13B"
                                },
                                "street": {
                                    "description": "Street",
                                    "type": "string",
                                    "example": "12, 14A"
                                },
                                "area": {
                                    "description": "Area",
                                    "type": "string",
                                    "example": "Salmiya, Sharq"
                                },
                                "building": {
                                    "description": "Building number",
                                    "type": "string",
                                    "example": "14, 13Z"
                                },
                                "details": {
                                    "description": "Any other address details",
                                    "type": "string",
                                    "example": "Al-Kuwait, near Hamra"
                                },
                                "notes": {
                                    "description": "Extra user notes",
                                    "type": "string",
                                    "example": "Do not park vehicle infront of the gate"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    }
                }
            }
        },
...
    },
...
}